Inventory Manager program
This project is a basic inventory manager program for a shoe retailer, created by me. The project is designed to exhibit a strong understanding of the foundations of Python concepts and software engineering, with a major focus around Object Oriented Programming and classes. The program reads the stock information from the 'inventory.txt' text file.